Added `TextureConfig`
The new `TextureConfig` structs specifies the parameters to create a texture.
It uses the builder pattern, and with its `build()` method can be used to
directly create a `Texture` while consuming the config.

Also added `Texture::from_raw_parts()` which allows to take control over the
Texture creation process (as per request of #23).