Added `RendererConfig`
The `RendererConfig` struct now contains the configurations for creating a `Renderer`.
With its `build()` it can also be used to directly create a `Renderer` while consuming the config.
Added the possibility to use custom shaders for the renderer (`RendererConfig::with_shaders()`)
Removed the already deprecated `Renderer::new_static()` method.

Added a whole bunch of new options in `TextureConfig` (these are now used in the `cube` example).
Fixed all compile errors from rebasing on the current commit.
In the `custom_textures` example, fixed the image using the wrong texture format.