Bump `wgpu` to 22.1
A few wgpu structs that this crate uses had fields added,
but they're ones where `Default::default` or `None` make sense,
so I've inserted them to get the build to work on the new version.

This is a breaking change for dependents relying on their own wgpu dependencies,
so I've bumped the version number and updated other dependencies that transitively
rely on wgpu.